refactor: change file name of profile to ss_profile
this file might be obselete when the next SoftSIM lib is released
