+++ lib/fs_port.c
@@ -7,7 +7,7 @@
 
 #include "ss_cache.h"
 #include "provision.h"
-#include "profile.h"
+#include "ss_profile.h"
 #include <onomondo/softsim/fs_port.h>
 #include <onomondo/softsim/list.h>
 #include <onomondo/softsim/utils.h>
